refactor: reorganize components and remove unused landing page files
