--- src/pages/LandingPage/components/footer.jsx
@@ -1,75 +1,0 @@
-import { Github, Linkedin } from "lucide-react";
-
-export default function Footer() {
-  return (
-    <footer
-      className={`relative lg:absolute lg:bottom-0 z-100 border-t border-border px-2 lg:px-6 w-full`}
-    >
-      <div className="py-6 p-2 md:px-14 lg:px-24 xl:px-56 mx-auto">
-
-        <div className="flex items-center justify-between">
-          {/* Copyright */}
-          <div>
-            <p className="text-primary/75 text-sm mb-0.5">
-              <a
-                href="https://opendata.lk"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                <span className="hover:text-accent">Open Data</span> ©{" "}
-              </a>
-              {new Date().getFullYear()}. All rights reserved.
-            </p>
-            <p className="text-primary/75 text-xs mb-0.5">
-              Licensed under{" "}
-              <a
-                href="https://creativecommons.org/licenses/by-nc-sa/4.0/"
-                target="_blank"
-                rel="noopener noreferrer"
-                className="text-accent hover:underline"
-              >
-                CC BY-NC-SA 4.0
-              </a>
-            </p>
-          </div>
-
-          {/* Social Media Links */}
-          <div className="flex items-center space-x-4">
-            <a
-              href="https://discord.gg/mg94NtHD9Y"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="p-2 text-primary/75 hover:text-accent transition-all duration-200 hover:scale-110"
-            >
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                width="20"
-                height="20"
-                viewBox="0 0 24 24"
-                fill="currentColor"
-              >
-                <path d="M20.317 4.37a19.791 19.791 0 0 0-4.885-1.515.074.074 0 0 0-.079.037c-.21.375-.444.864-.608 1.25a18.27 18.27 0 0 0-5.487 0 12.64 12.64 0 0 0-.617-1.25.077.077 0 0 0-.079-.037A19.736 19.736 0 0 0 3.677 4.37a.07.07 0 0 0-.032.027C.533 9.046-.32 13.58.099 18.057a.082.082 0 0 0 .031.057 19.9 19.9 0 0 0 5.993 3.03.078.078 0 0 0 .084-.028c.462-.63.874-1.295 1.226-1.994a.076.076 0 0 0-.041-.106 13.107 13.107 0 0 1-1.872-.892.077.077 0 0 1-.008-.128 10.2 10.2 0 0 0 .372-.292.074.074 0 0 1 .077-.01c3.928 1.793 8.18 1.793 12.062 0a.074.074 0 0 1 .078.01c.12.098.246.198.373.292a.077.077 0 0 1-.006.127 12.299 12.299 0 0 1-1.873.892.077.077 0 0 0-.041.107c.36.698.772 1.362 1.225 1.993a.076.076 0 0 0 .084.028 19.839 19.839 0 0 0 6.002-3.03.077.077 0 0 0 .032-.054c.5-5.177-.838-9.674-3.549-13.66a.061.061 0 0 0-.031-.03zM8.02 15.33c-1.183 0-2.157-1.085-2.157-2.419 0-1.333.956-2.419 2.157-2.419 1.21 0 2.176 1.096 2.157 2.42 0 1.333-.956 2.418-2.157 2.418zm7.975 0c-1.183 0-2.157-1.085-2.157-2.419 0-1.333.955-2.419 2.157-2.419 1.21 0 2.176 1.096 2.157 2.42 0 1.333-.946 2.418-2.157 2.418z" />
-              </svg>
-            </a>
-            <a
-              href="https://www.linkedin.com/company/lankadata/"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="p-2 text-primary/75 hover:text-accent transition-all hover:scale-110"
-            >
-              <Linkedin className="w-5 h-5" />
-            </a>
-            <a
-              href="https://github.com/LDFLK"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="p-2 text-primary/75 hover:text-accent transition-all duration-200 hover:scale-110"
-            >
-              <Github className="w-5 h-5" />
-            </a>
-          </div>
-        </div>
-      </div>
-    </footer >
-  );
-}
